crewai_agent_app_sample_code/requirements.txt
crewai[tools]
pydantic>=2.9.2
databricks-agents
mlflow>=2.18.0
databricks-sdk[openai]
databricks-vectorsearch
pyyaml
git+https://github.com/unitycatalog/unitycatalog.git#subdirectory=ai/core
git+https://github.com/unitycatalog/unitycatalog.git#subdirectory=ai/integrations/crewai
tabulate
pandas
pyspark
databricks-connect==15.1.0